post_service/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
pydantic
pydantic_settings
psycopg2-binary
protobuf
SQLAlchemy==2.0.38
aiokafka
